WIP: default to plain TCP
I'll have to figure out a mechanism for handing the preference for TLS
down to the connect function.

For now I'll default to plain TCP to preserve existing functionality.

It doesn't obviously fail with the consumer or producer examples, but
they also don't appear to work correctly since I don't see message flow.