Add option to bypass TLS Certificate validation
 - Add TLSTrusted and TLSUntrusted variants of TLS setup

The default remains to use unsecured TCP connections where there is no
reasonable expectation of confidentiality or integrity.

TLSTrusted is the recommended setting where the AMQP servers support
it.  Using it will require that the AMQP servers identify themselves
with trusted certificates.  Private CAs and self-signed certificates can
be supported using mechanisms dependent on the trust store of the
underlying Operating System. This is out of the scope of this package to
explain, but examples are readily available for most Operating Systems.

TLSUntrusted is an insecure option and should never be used in
production.  Bypassing TLS certificate validation opens client code and
its users to man-in-the-middle attacks.  As I have described above it is
relatively straight-forward to trust a self-signed cert or a private CA, so there
is very rarely a valid use case for TLSUntrusted.